pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coderag"
version = "0.1.0"
description = "Lean proof-of-concept retrieval-augmented assistant for codebases."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE-2.0.txt" }
authors = [{ name = "Neverdecel" }]
dependencies = [
    "faiss-cpu>=1.9.0.post1,<2",
    "numpy>=1.26,<2",
    "openai>=1.44,<2",
    "python-dotenv>=1.0,<2",
    "sentence-transformers>=3.0,<4",
    "streamlit>=1.38,<2",
    "tenacity>=8.5,<9",
    "watchdog>=4.0,<5",
]

[project.optional-dependencies]
dev = [
    "black>=24.8",
    "flake8>=7.1",
    "isort>=5.13",
    "mypy>=1.11",
    "pre-commit>=3.7",
    "pytest>=8.3",
]

[project.scripts]
coderag-cli = "coderag.cli:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    # directories
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | env
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["coderag"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = false
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = true
